realised work with TextField

+++ BackgroundSettingAppSwiftUI/Views/BGSettingContentView.swift
@@ -8,9 +8,9 @@
 import SwiftUI
 
 struct BGSettingContentView: View {
-    @State private var redSliderValue = Double.random(in: 0...255)
-    @State private var greenSliderValue = Double.random(in: 0...255)
-    @State private var blueSliderValue = Double.random(in: 0...255)
+    @State private var redSliderValue = Double(Int.random(in: 0...255))
+    @State private var greenSliderValue = Double(Int.random(in: 0...255))
+    @State private var blueSliderValue = Double(Int.random(in: 0...255))
     
     var body: some View {
         ZStack {
